+++ CHANGES.rst
@@ -5,8 +5,10 @@ Changelog
 0.20.1 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- Avoid error in `creating_group_split` when a group is not correct.
+  [sgeulette]
+- Taken into account routing table in `IncomingEmail` consumer.
+  [sgeulette]
 
 0.20.0 (2023-11-28)
 -------------------